Prevent rebuilds when a dependency is sometimes only needed for host builds.
I was observing rebuilds when a dependency is sometimes host-only and other times built for target as well.

Cargo seems to have some optimization that would unify these when the target build is present, causing
a different context and corresponding hash.

Because this optimization only triggers when the target build is present, we need to output to the target list so that this optimization triggers every time.